+++ prompt_styles/human.sh
@@ -3,46 +3,34 @@
 # Author: Diogo Alexsander Cavilha <diogocavilha@gmail.com>
 # Date:   06.11.2018
 
-. ~/.fancy-git/aliases
-. ~/.fancy-git/fancygit-completion
-. ~/.fancy-git/commands.sh
-
 # ----------------------------------------------------------------------------------------------------------------------
 # The main function to change the prompt.
 # ----------------------------------------------------------------------------------------------------------------------
 fancygit_prompt_builder() {
-    . ~/.fancy-git/config.sh
     . ~/.fancy-git/modules/update-manager.sh
 
     check_for_update
-    
+
     # !! IMPORTANT !!
-    # If you're just interested on creating a new color scheme, here you have the skin color config.
-    # Once you wish to change some colors, the only thing you need to to is changing these *_foreground variable values.
+    # If you're just interested on creating a new color scheme, check $HOME/.fancy-git/color_schemes.
     # It'll be handled in order to create the proper color on PS1 prompt.
-    local time_foreground="231"
-    local user_symbol_foreground="231"
-    local workdir_foreground="10"
-    local user_foreground="208"
-    local host_foreground="220"
-    local at_foreground="231"
-    local branch_foreround="13"
-    local branch_staged_files_foreground="13"
-    local branch_changed_files_foreground="13"
+    local color_scheme_name
+    color_scheme_name=$(fancygit_config_get "style" "default")
+    . "$HOME/.fancy-git/color_schemes/$color_scheme_name"
 
     # !! WARNING !!
     # From here you better now what you're doing. Have fun =D
 
     # Create color tags to change prompt style.
-    local time_color_tag="\\[\\e[38;5;${time_foreground}m\\]"
-    local user_color_font_tag="\\[\\e[38;5;${user_foreground}m\\]"
-    local host_color_font_tag="\\[\\e[38;5;${host_foreground}m\\]"
-    local at_color_font_tag="\\[\\e[38;5;${at_foreground}m\\]"
-    local user_symbol_color_font_tag="\\[\\e[38;5;${user_symbol_foreground}m\\]"
-    local workdir_color_font_tag="\\[\\e[38;5;${workdir_foreground}m\\]"
-    local branch_color_staged_files_font_tag="\\[\\e[38;5;${branch_staged_files_foreground}m\\]"
-    local branch_color_changed_files_font_tag="\\[\\e[38;5;${branch_changed_files_foreground}m\\]"
-    local branch_color_font_tag="\\[\\e[38;5;${branch_foreround}m\\]"
+    local time_color_tag="\\[\\e[38;5;${fancygit_color_scheme_time_foreground}m\\]"
+    local user_color_font_tag="\\[\\e[38;5;${fancygit_color_scheme_user_foreground}m\\]"
+    local host_color_font_tag="\\[\\e[38;5;${fancygit_color_scheme_host_foreground}m\\]"
+    local at_color_font_tag="\\[\\e[38;5;${fancygit_color_scheme_at_foreground}m\\]"
+    local user_symbol_color_font_tag="\\[\\e[38;5;${fancygit_color_scheme_user_symbol_foreground}m\\]"
+    local workdir_color_font_tag="\\[\\e[38;5;${fancygit_color_scheme_workdir_foreground}m\\]"
+    local branch_color_staged_files_font_tag="\\[\\e[38;5;${fancygit_color_scheme_branch_staged_files_foreground}m\\]"
+    local branch_color_changed_files_font_tag="\\[\\e[38;5;${fancygit_color_scheme_branch_changed_files_foreground}m\\]"
+    local branch_color_font_tag="\\[\\e[38;5;${fancygit_color_scheme_branch_foreround}m\\]"
     local color_reset="\\[\\e[39m\\]"
     local bold="\\[\\e[1m\\]"
     local bold_reset="\\[\\e[0m\\]"
@@ -68,14 +56,14 @@ fancygit_prompt_builder() {
     local prompt_time=""
     local fancygit_PS2=""
     local is_double_line=""
-    local notification_area=""
     local branch_status=""
+    local is_rich_notification
 
     # Get git repo info.
-    branch_status=$(__get_git_status)
-    staged_files=$(__get_git_diff)
+    branch_status=$(fancygit_git_get_status)
+    staged_files=$(fancygit_git_get_staged_files)
 
-    if [ "${branch_status}" != "" ]
+    if [ "$branch_status" != "" ]
     then
         branch="${branch_color_changed_files_font_tag}${bold}"
         branch_end="${background_reset}${color_reset}"
@@ -88,6 +76,7 @@ fancygit_prompt_builder() {
     fi
 
     # Check some config preferences.
+    is_rich_notification=$(fancygit_config_get "show-rich-notification" "true")
     if fancygit_config_is "double-line" "true"
     then
         fancygit_PS2=$(fancygit_config_get "ps2" "➜")
@@ -96,8 +85,8 @@ fancygit_prompt_builder() {
 
     if fancygit_config_is "show-time" "true"
     then
-      formatted_time=$(date +"${time_format}")
-      prompt_time="${time}[${formatted_time}]${time_end} "
+        time_format=$(fancygit_config_get "time-format" "%H:%M:%S")
+        prompt_time="${time}[$(date +"$time_format")] ${time_end}"
     fi
 
     if fancygit_config_is "show-user-at-machine" "true"
@@ -112,19 +101,16 @@ fancygit_prompt_builder() {
     fi
 
     prompt_symbol="${user_symbol}\$${user_symbol_end}"
-
-    if ! [ -z ${VIRTUAL_ENV} ] || ([ "${CONDA_DEFAULT_ENV}" != "base" ] && ! [ -z ${CONDA_DEFAULT_ENV} ])
-    then
-        venv="${working_on_venv}"
-    fi
-
+    venv=$(__fancygit_get_venv_icon)
     prompt_path="${path}${venv}${path_sign}${path_end}${color_reset}"
 
+    # If we have a branch name, it means we are in a git repo, so we need to make some changes on PS1.
+    branch_name=$(fancygit_git_get_branch)
     if [ "${branch_name}" != "" ]
     then
         prompt_path="${path_git}${path_sign}${path_end}"
         prompt_branch="${branch}${branch_name}${branch_end}"
-        PS1="${bold}${prompt_time}${prompt_user}${prompt_path} on ${prompt_branch}$(__fancygit_notification_area 1)"
+        PS1="${bold}${prompt_time}${prompt_user}${prompt_path} on ${prompt_branch}$(__fancygit_get_notification_area "$is_rich_notification")"
         PS1="${PS1}${prompt_symbol}${is_double_line}${bold_reset} "
         return
     fi
@@ -133,178 +119,22 @@ fancygit_prompt_builder() {
 }
 
 # ----------------------------------------------------------------------------------------------------------------------
-# Get the notification area according to the parameter.
-# 
-# param int 0: Poor notification area, no icons.
-# param int 1: Rich notification area, icons instead of simples characters.
-# return string Notification area.
+# Return the icon that indicates we're working in a virtual environment.
 # ----------------------------------------------------------------------------------------------------------------------
-__fancygit_notification_area() {
-    if [ 0 -eq ${1} ]
+__fancygit_get_venv_icon() {
+    local icon_venv
+
+    icon_venv=" "
+
+    if ! [ -z $VIRTUAL_ENV ] || ([ "$CONDA_DEFAULT_ENV" != "base" ] && ! [ -z $CONDA_DEFAULT_ENV ])
     then
-        echo " $(__fancygit_poor_notification_area) "
-        return
-    fi
-
-    echo " $(__fancygit_rich_notification_area) "
-}
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Creates a rich notification area, which means it will have the same icons as fancy themes.
-# ----------------------------------------------------------------------------------------------------------------------
-__fancygit_rich_notification_area() {
-    local notification_area=""
-    local has_added_files=""
-    local has_git_stash=""
-    local has_untracked_files=""
-    local has_changed_files=""
-    local has_unpushed_commits=""
-
-    has_git_stash=" "
-    has_untracked_files="  "
-    has_changed_files="  "
-    has_added_files="  "
-    has_unpushed_commits="  "
-    working_on_venv=" "
-
-    if [ "${staged_files}" = "" ]
-    then
-        has_added_files=""
-    fi
-
-    if [ "${git_stash}" = "" ]
-    then
-        has_git_stash=""
-    fi
-
-    if [ "${git_untracked_files}" = "" ]
-    then
-        has_untracked_files=""
-    fi
-
-    if [ "${git_changed_files}" = "" ]
-    then
-        has_changed_files=""
-    fi
-
-    has_unpushed_commits="${has_unpushed_commits}+${git_number_unpushed_commits}"
-    if [ "${git_has_unpushed_commits}" = "" ]
-    then
-        has_unpushed_commits=""
-    fi
-
-    notification_area="${has_git_stash}${has_untracked_files}${has_changed_files}${has_added_files}${has_unpushed_commits}"
-
-    echo "${notification_area}"
-}
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Creates a poor notification area, which means it won't have icons.
-# ----------------------------------------------------------------------------------------------------------------------
-__fancygit_poor_notification_area() {
-    local notification_area=""
-    local color_reset=""
-    local color_cyan=""
-    local color_light_green=""
-    local color_light_red=""
-
-    color_reset="\\[\\e[39m\\]"
-    color_cyan="\\[\\e[36m\\]"
-    color_light_green="\\[\\e[92m\\]"
-    color_light_red="\\[\\e[91m\\]"
-
-    # Git info
-    local remote_name=""
-    local remote_name=""
-    local branch_name=""
-    local staged_files=""
-    local git_stash=""
-    local git_untracked_files=""
-    local git_changed_files=""
-    local git_has_unpushed_commits=""
-    local git_number_unpushed_commits=""
-    local git_number_untracked_files=""
-    local git_number_changed_files=""
-    local merged_branch=""
-
-    remote_name=$(git rev-parse --abbrev-ref --symbolic-full-name @{u} 2> /dev/null | cut -d"/" -f1)
-    remote_name=${remote_name:-origin}
-    branch_name=$(git rev-parse --abbrev-ref HEAD 2> /dev/null)
-    staged_files=$(__get_git_diff)
-    git_stash=$(git stash list 2> /dev/null)
-    git_untracked_files=$(git ls-files --others --exclude-standard 2> /dev/null)
-    git_changed_files=$(git ls-files -m 2> /dev/null)
-    git_has_unpushed_commits=$(git log ${remote_name}/${branch_name}..HEAD 2> /dev/null)
-    git_number_unpushed_commits=$(git log --pretty=oneline ${remote_name}/${branch_name}..HEAD 2> /dev/null | wc -l)
-    git_number_untracked_files=$(git ls-files --others --exclude-standard 2> /dev/null | wc -w)
-    git_number_changed_files=$(git ls-files -m 2> /dev/null | wc -l)
-    merged_branch=""
-
-    if [ "$branch_name" != "master" ]; then
-        merged_branch=$(git branch -r --merged master 2> /dev/null | grep ${branch_name} 2> /dev/null)
-    fi
-
-    if [ "${git_stash}" != "" ]
-    then
-        notification_area="${notification_area}∿${color_reset} "
-    fi
-
-    if [ "${git_number_untracked_files}" -gt 0 ]
-    then
-        notification_area="${notification_area}${color_cyan}?${color_reset} "
-    fi
-
-    if [ "${git_number_changed_files}" -gt 0 ]
-    then
-        notification_area="${notification_area}${color_light_green}+${color_reset}${color_light_red}-${color_reset} "
-    fi
-
-    if [ "${staged_files}" != "" ]
-    then
-        notification_area="${notification_area}${color_light_green}✔${color_reset} "
-    fi
-
-    if [ "${git_has_unpushed_commits}" ]
-    then
-        notification_area="${notification_area}${color_light_green}^${git_number_unpushed_commits}${color_reset} "
-    fi
-
-    if [ "${branch_name}" != "" ] && __fancygit_is_only_local_branch "${branch_name}"
-    then
-        notification_area="${notification_area}${color_light_green}*${color_reset} "
-    fi
-
-    if [ "${merged_branch}" != "" ]
-    then
-        notification_area="${notification_area}${color_light_green}<${color_reset} "
-    fi
-
-    if [ "${notification_area}" != "" ]
-    then
-        # Trim notification_area content
-        notification_area=$(echo "${notification_area}" | sed -e 's/[[:space:]]*$//' | sed -e 's/^[[:space:]]*//')
-        echo "[${notification_area}]"
+        echo "$icon_venv"
         return
     fi
 
     echo ""
 }
 
-# Git commands.
-# Some git commands are used in many parts of the code, that's why we have written some of them into a function.
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Performs a git status and return the output.
-# ----------------------------------------------------------------------------------------------------------------------
-__get_git_status() {
-    git status -s 2> /dev/null
-}
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Performs a git diff and return the output.
-# ----------------------------------------------------------------------------------------------------------------------
-__get_git_diff() {
-    git diff --name-only --cached 2> /dev/null
-}
-
+# Here's where the magic happens!
+# It calls our main function (fancygit_prompt_builder) in order to mount a beautiful PS1 prompt =D
 PROMPT_COMMAND="fancygit_prompt_builder"